# arthurclive_swagger.yaml
swagger: '2.0'
info:
  description: Arthur Clive Web Api
  version: 1.0.0
  title: Arthur Clive Swagger
  contact:
    email: raguvarthan.n@turingminds.com
host: arhturclive.swagger.io
basePath: /v1
...
schemes:
  - http
paths:
...
  /api/user/userinfo:
    post:
      tags:
        - User
      operationId: postUserinfo
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: body
          name: data
          description: Object that need to be added to the DB
          required: true
          schema:
            $ref: '#/definitions/UserInfo'
      responses:
        '200':
          description: Inserted
          schema:
            $ref: '#/definitions/ResponseData'
        '201':
          description: User Updated + responce
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
        '401':
          description: Billing or Shipping Address Not Avaliable
          schema:
            $ref: '#/definitions/ResponseData'
    get:
      tags:
        - User
      operationId: getAllUsers
      produces:
        - application/xml
        - application/json
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserInfoDetails'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
  '/api.user/userinfo/{phonenumber}':
    get:
      tags:
        - User
      operationId: getOneUser
      produces:
        - application/xml
        - application/json
      parameters:
        - name: phonenumber
          in: path
          description: Phonenumber of user whoes details need to be filtered.
          required: true
          type: string
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/UserInfoDetails'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
  '/api/user/cart/{username}':
    post:
      tags:
        - Cart
      operationId: insertProductToCart
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: path
          name: username
          description: Username of the user
          required: true
          type: string
        - in: body
          name: cartProduct
          description: Object that need to be added to the DB
          required: true
          schema:
            $ref: '#/definitions/CartList'
      responses:
        '200':
          description: Inserted
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
    get:
      tags:
        - Cart
      summary: Get all the products in cart
      description: ''
      operationId: getProductsInCart
      produces:
        - application/xml
        - application/json
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/CartList'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
  '/api/user/wishlist/{username}':
    post:
      tags:
        - WishList
      operationId: insertProductToWishList
      consumes:
        - application/json
        - application/xml
      produces:
        - application/xml
        - application/json
      parameters:
        - in: path
          name: username
          description: Username of the user
          required: true
          type: string
        - in: body
          name: wishlistProduct
          description: Object that need to be added to the DB
          required: true
          schema:
            $ref: '#/definitions/WishlistList'
      responses:
        '200':
          description: Inserted
          schema:
            $ref: '#/definitions/ResponseData'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
    get:
      tags:
        - WishList
      operationId: getProductsInWishlist
      produces:
        - application/xml
        - application/json
      responses:
        '200':
          description: Success
          schema:
            $ref: '#/definitions/WishlistList'
        '400':
          description: Failed
          schema:
            $ref: '#/definitions/ResponseData'
...
definitions:
  ResponseData:
    type: object
    properties:
      Code:
        type: number
      Message:
        type: string
      Data:
        type: object
      Content:
        type: object
...
  CartDetails:
    type: object
    properties:
      Id:
        type: number
      UserName:
        type: string
      ProductSKU:
        type: string
      MinioObject_Url:
        type: string
      ProductFor:
        type: string
      ProductType:
        type: string
      ProductDesign:
        type: string
      ProductBrand:
        type: string
      ProductPrice:
        type: number
      ProductDiscount:
        type: number
      ProductDiscountPrice:
        type: number
      ProductQuantity:
        type: number
      ProdyctSize:
        type: string
      ProductColour:
        type: string
      ProductDescription:
        type: string
  WishListDetails:
    type: object
    properties:
      Id:
        type: number
      UserName:
        type: string
      ProductSKU:
        type: string
      MinioObject_Url:
        type: string
      ProductFor:
        type: string
      ProductType:
        type: string
      ProductDesign:
        type: string
      ProductBrand:
        type: string
      ProductPrice:
        type: number
      ProductDiscount:
        type: number
      ProductDiscountPrice:
        type: number
      ProdyctSize:
        type: string
      ProductColour:
        type: string
      ProductDescription:
        type: string
  UserInfo:
    type: object
    properties:
      PhoneNumber:
        type: number
  UserInfoDetails:
    type: object
    properties:
      Id:
        type: number
      PhoneNumber:
        type: number
      FullName:
        type: string
      Email:
        type: string
      ShippingAddress:
        type: array
        items:
          $ref: '#/definitions/Address'
      BillingsAddress:
        type: array
        items:
          $ref: '#/definitions/Address'
  Address:
    type: object
    properties:
      AddressId:
        type: integer
      PhoneNumber:
        type: number
      Default:
        type: boolean
      DoorNo_Or_BuildingName:
        type: string
      Area_Or_Street:
        type: string
      City:
        type: string
      State:
        type: string
      Pincode:
        type: string
      Landmark:
        type: string
  CartList:
    type: object
    properties:
      ListOfProducts:
        type: array
        items:
          $ref: '#/definitions/CartDetails'
  WishlistList:
    type: object
    properties:
      ListOfProducts:
        type: array
        items:
          $ref: '#/definitions/WishListDetails'
...